+++ library/src/main/java/it/sephiroth/android/library/tooltip/TooltipOverlay.java
@@ -1,10 +1,13 @@
 package it.sephiroth.android.library.tooltip;
 
 import android.content.Context;
+import android.content.res.TypedArray;
 import android.util.AttributeSet;
 import android.widget.ImageView;
 
 public class TooltipOverlay extends ImageView {
+    private int mMargins;
+
     public TooltipOverlay (Context context) {
         this(context, null);
     }
@@ -21,10 +24,20 @@ public class TooltipOverlay extends ImageView {
     private void init (final Context context, final int defStyleResId) {
         TooltipOverlayDrawable drawable = new TooltipOverlayDrawable(context, defStyleResId);
         setImageDrawable(drawable);
+
+        final TypedArray array =
+            context.getTheme().obtainStyledAttributes(defStyleResId, R.styleable.TooltipOverlay);
+        mMargins = array.getDimensionPixelSize(R.styleable.TooltipOverlay_android_layout_margin, 0);
+        array.recycle();
+
     }
 
     public TooltipOverlay (Context context, AttributeSet attrs, int defStyleAttr, int defStyleResId) {
         super(context, attrs, defStyleAttr);
         init(context, defStyleResId);
+    }
+
+    public int getLayoutMargins () {
+        return mMargins;
     }
 }